Update mod_publication (student folders) to r3.2.2

+++ mod/publication/lib.php
@@ -15,12 +15,10 @@
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
 /**
- * lib.php
+ * Contains interface and callback methods for mod_publication
  *
  * @package       mod_publication
- * @author        Andreas Hruska (andreas.hruska@tuwien.ac.at)
- * @author        Katarzyna Potocka (katarzyna.potocka@tuwien.ac.at)
- * @author        Philipp Hager (office@phager.at)
+ * @author        Philipp Hager
  * @author        Andreas Windbichler
  * @copyright     2014 Academic Moodle Cooperation {@link http://www.academic-moodle-cooperation.org}
  * @license       http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
@@ -32,9 +30,8 @@ defined('MOODLE_INTERNAL') || die();
  * Adds a new publication instance
  *
  * @param publication $publication
- * @param mform $mform
  */
-function publication_add_instance($publication, $mform = null) {
+function publication_add_instance($publication) {
     global $DB, $OUTPUT;
 
     $cmid       = $publication->coursemodule;
@@ -103,9 +100,8 @@ function publication_supports($feature) {
  * updates an existing publication instance
  *
  * @param publication $publication
- * @param mform $mform
  */
-function publication_update_instance($publication, $mform = null) {
+function publication_update_instance($publication) {
     global $DB;
 
     $publication->id = $publication->instance;
@@ -221,8 +217,9 @@ function publication_reset_userdata($data) {
 
             $fs = get_file_storage();
             foreach ($filerecords as $filerecord) {
-                $file = $fs->get_file_by_id($filerecord->fileid);
-                $file->delete();
+                if ($file = $fs->get_file_by_id($filerecord->fileid)) {
+                    $file->delete();
+                }
             }
 
             $DB->delete_records('publication_file', array('publication' => $publication->id));
